+++ tests/__files__/merge/vehicle.test2.pak
