+++ tests/sync/extract.test.ts
@@ -1,8 +1,8 @@
 import { pathExistsSync, removeSync } from 'fs-extra';
-import Makeobj from '../src/Makeobj';
+import Makeobj from '../../src/Makeobj';
 
 test('extract', () => {
-  const pakfile = './tests/extract/merged.pak';
+  const pakfile = './tests/__files__/extract/merged.pak';
   expect(pathExistsSync(pakfile)).toBeTruthy();
 
   const makeobj = new Makeobj(process.env.MAKEOBJ_PATH);
@@ -10,8 +10,8 @@ test('extract', () => {
   expect(result.isSuccess).toBeTruthy();
 
   const shouldExtractFiles = [
-    './tests/extract/citycar.test1.pak',
-    './tests/extract/vehicle.test2.pak',
+    './tests/__files__/extract/citycar.test1.pak',
+    './tests/__files__/extract/vehicle.test2.pak',
   ];
 
   shouldExtractFiles.map(file => {
