+++ tests/sync/list.test.ts
@@ -1,8 +1,8 @@
 import { pathExistsSync } from 'fs-extra';
-import Makeobj from '../src/Makeobj';
+import Makeobj from '../../src/Makeobj';
 
 test('list', () => {
-  const pakfile = './tests/list/test.pak';
+  const pakfile = './tests/__files__/list/test.pak';
   expect(pathExistsSync(pakfile)).toBeTruthy();
 
   const makeobj = new Makeobj(process.env.MAKEOBJ_PATH);
